backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
Deprecated==1.2.18
deprecation==2.1.0
exceptiongroup==1.3.1
fastapi==0.128.0
flupy==1.2.3
fsspec==2026.1.0
greenlet==3.3.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.7
langchain-text-splitters==1.1.0
langsmith==0.6.7
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.0
multidict==6.7.1
numpy==1.26.4
ollama==0.6.1
orjson==3.11.6
packaging==25.0
pgvector==0.3.6
postgrest==2.27.2
propcache==0.4.1
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.11.0
PyMuPDF==1.26.7
pyparsing==3.3.2
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
PyYAML==6.0.3
realtime==2.27.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.1
six==1.17.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.46
starlette==0.50.0
storage3==2.27.2
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.27.2
supabase-auth==2.27.2
supabase-functions==2.27.2
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.0
uvicorn==0.40.0
vecs==0.4.5
websockets==15.0.1
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zstandard==0.25.0
